auto configure to use extracted libraries in graal native image
