+++ jme3-nativeimage-plugin/src/main/java/org/jmonkeyengine/gradle/nativemetadata/internal/JmeNativeRuntimeDefaultsHook.java
@@ -1,0 +1,11 @@
+package org.jmonkeyengine.gradle.nativemetadata.internal;
+
+import com.oracle.svm.core.jdk.RuntimeSupport;
+
+final class JmeNativeRuntimeDefaultsHook implements RuntimeSupport.Hook {
+
+    @Override
+    public void execute(boolean firstIsolate) {
+        JmeNativeRuntimeDefaults.apply();
+    }
+}
